add some todos, migrate some stuff to util.js
