+++ README.md
@@ -14,14 +14,15 @@
 
 ## 🌈 Features
 
-* 🎉 Easy to install, run and update
-* ✉️ Update checker and log notifications
+* 🎉 [Easy to install, run and update](#️-how-to-use-this)
+* ✉️ [Update checker and log notifications](#-how-to-update-the-script)
 * 👽 Same logic as the [PHP version](https://github.com/SteamDatabase/SalienCheat) (we almost have parity)
-* 👌 Pick your own steam group
-* 👥 Works well with multiple tokens/scripts
-* 👀 Name your running scripts
-* 🐳 Docker support
-* ☁️ Heroku support
+* 👌 [Pick your own steam group](#-represent-your-steam-group-optional)
+* 👥 [Works well with multiple tokens/scripts](#-multiple-tokensscripts)
+* 👀 [Name your running scripts](#-multiple-tokensscripts)
+* 📦 [npm package export](#advanced--usage-as-an-npm-package)
+* 🐳 [Docker support](#advanced--running-as-a-docker-container)
+* ☁️ [Heroku support](#advanced-️-deploying-to-heroku)
 
 > Note: We'll try our best to keep this version up to date with the PHP and other versions! Suggestions welcome.
 
@@ -64,8 +65,8 @@ You can get your group id by going to https://steamcommunity.com/groups/YOUR_GRO
 If you'd like to team up with an established larger group please consider using either:
 
 * [/r/saliens](https://steamcommunity.com/groups/summersaliens) id: `103582791462557324`
-* [100Pals](https://steamcommunity.com/groups/100pals) id: `103582791454524084`
 * [SteamDB](https://steamcommunity.com/groups/steamdb) id: `103582791434298690`
+* [100Pals](https://steamcommunity.com/groups/100pals) id: `103582791454524084`
 
 ### 👥 Multiple tokens/scripts
