+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "salien-script-js",
-  "version": "0.0.11",
+  "version": "0.0.12",
   "description": "A easy to install, run and update Node.js script for the Steam salien mini-game.",
   "keywords": [
     "salien",
